refactor[web-35]: replace individual input components with input forms
