--- src/components/UiKit/FormInputs/ResidencyInfo/residencyInfo.module.scss
@@ -1,19 +1,0 @@
-.container {
-  contain: layout;
-  box-sizing: border-box;
-  width: 350px;
-
-  .title {
-    font-size: 16px;
-    margin-bottom: 10px;
-    color: #000000;
-  }
-}
-
-.error_message {
-  bottom: -18px;
-  min-height: 15px;
-  margin-top: 3px;
-  font-size: 0.765rem;
-  color: var(--common-color-text-text-error);
-}
